+++ questShared/src/commonMain/kotlin/ru/neura/questgoshared/questGoShared/data/features/configuration/models/ConfigurationModel.kt
@@ -1,7 +1,7 @@
 package ru.neura.questgoshared.questGoShared.data.features.configuration.models
 
 data class ConfigurationModel(
-    val currentQuestId: Long,
-    val currentQuestPage: Long,
-    val boughtQuestIds: List<Long>
+    val currentQuestId: Int,
+    val currentQuestPage: Int,
+    val boughtQuestIds: List<Int>
 )
